+++ src/transcriber_service/services/__init__.py
@@ -1,4 +1,4 @@
-from .processing import WhisperProcessor
+from .transcribe_processor import WhisperProcessor
 from .storage_service import StorageService
 from .auth_service import AuthService
 from .audio_service import AudioService
