+++ src/transcriber_service/services/transcribe_processor.py
